[ENH] Encode column data type in output dictionary (#220)
* [REF] Minor fixes in the e2e test

* [REF] Factor out data dictionary code

Create the dictionary separately in a utility
function.

* [REF] Create custom hook for schema validation

Also move data dictionary hook into hooks
module

* Add mapping from variabletype to bids

* Updating interfaces with variableType

* Revert type changes

* [REF] dataType to bidsType for Columns

* Add missing update in the mocks

* Add VariableType to types

* Integrating column variabletype

* Refactor config dataType -> variableType

This is only working because of the fallback
to the local config - where we have already
updated the types to use variableType.

We will have to update the communities repo as
well to make this work.

* change over to storing variabletype

* Add BIDS Type

* Start writing out variableType

Still requires updated config files
And we need to also update the conditions
for what components to show and where

(i.e. the multi-column thing)

* Basic update for column filtering

* Updated reliance on is_multicolumn

* Load variabletype from data dictionary

* Update variable type in mocks

* Fix some type errors in the tests

* Refactor the multi-column state getter

* Consolidate to one variableType

* Revert "Updated reliance on is_multicolumn"

This reverts commit 4200b7d5b28d90a6916c7a832a329b402413b0d4.
We want the matching config standardized Variable back

* [FIX] Do not show participantID on value annotation

* Fix typos

And also remove breaking character
in URL

* Remove participantID from mappedSV getter mock

* Apply suggestions from code review

* Apply more suggestions from code review

Thanks Arman